Fixed client to fetch objects not hash maps.

+++ src/main/kotlin/com/djdapz/bitcoin_graph/hit_btc/domain/HitBtcSymbol.kt
@@ -4,7 +4,7 @@ data class HitBtcSymbol(
         val id: String,
         val baseCurrency: String,
         val quoteCurrency: String,
-        val quantityIncrement: Int,
+        val quantityIncrement: Double,
         val tickSize: Double,
         val takeLiquidityRate: Double,
         val provideLiquidityRate: Double,
